Suppress gymnasium and pygame warnings at startup
- Use explicit float32 dtype in action space arrays to prevent precision warnings
- Add warnings.filterwarnings() to suppress gymnasium, pygame, and pkg_resources warnings

+++ src/drone_ai/environment.py
@@ -111,15 +111,15 @@ class DroneEnv(gym.Env):
         if task in [TaskType.DELIVERY, TaskType.DELIVERY_ROUTE]:
             # 4 motor commands [-1, 1] + 1 drop signal [0, 1]
             self.action_space = spaces.Box(
-                low=np.array([-1.0, -1.0, -1.0, -1.0, 0.0]),
-                high=np.array([1.0, 1.0, 1.0, 1.0, 1.0]),
+                low=np.array([-1.0, -1.0, -1.0, -1.0, 0.0], dtype=np.float32),
+                high=np.array([1.0, 1.0, 1.0, 1.0, 1.0], dtype=np.float32),
                 dtype=np.float32
             )
         else:
             # 4 motor commands [-1, 1] for reversible fans
             self.action_space = spaces.Box(
-                low=-1.0,
-                high=1.0,
+                low=np.float32(-1.0),
+                high=np.float32(1.0),
                 shape=(4,),
                 dtype=np.float32
             )
